fixing sleep in test by adding retries; updating tests to use TSO timestamps
added retry tso client

Make ExpBackoff pause between executions

update

update test

use "cpo" for cpo cmdline config in tests

changed tsoClient to ptr

update test3si timestamps

fix test

fix test finalizing wrong txn

changed all timestamps to get from tso

update async finalization wait time

+++ src/k2/transport/RetryStrategy.cpp
@@ -33,7 +33,7 @@ ExponentialBackoffStrategy& ExponentialBackoffStrategy::withRetries(int retries)
 }
 
 // Set the exponential increase rate
-ExponentialBackoffStrategy& ExponentialBackoffStrategy::withRate(int rate) {
+ExponentialBackoffStrategy& ExponentialBackoffStrategy::withRate(double rate) {
     K2LOG_D(log::tx, "rate: {}", rate);
     _rate = rate;
     return *this;
